update ts-jest and ts-jest config so that we run tests as native ESM in Node, remove and replace microbundle build with tsc, update tsconfigs so that we output only JS modules (ESM), add a dev script for building in watch mode, add DOM lib for missing WebAssembly types in test file, remove unnecessary stuff from .gitignore, use only main and types package.json fields to specify the main export and types (all other imports into dist/ will have type definitions as expected), update README
This change makes the project easily consumable in the vast majority of TypeScript projects because it follows the simplest possible project and output layout. In this format, most projects, tools, and even browser applications, will have the easiest time trying to import the library with the least amount of configuration, including with type support.

By dropping microbundle we also dropped support for UMD, CommonJS, and multiple variations of ES modules. It is better to start with something simple that works, then if we want a UMD build we can add that, but we definitely don't need a CJS build. On with the ESM future!

BREAKING: `import { w } from 'wazum'` no longer works. The minimal upgrade path is to use `import * as w from 'wazum'`. The new format is friendlier for anyone who has ESM-based tree shaking in their build setup; importing in the form `import { Module, add } from 'wazum'` will have better tree shaking support with today's tools.